jslocalstorageexample

+++ Example/localstorageDataInTableScript_5.js
@@ -6,18 +6,18 @@ var userArray = [];
 var userDetailArray = [];
 var userArrayString;
 var dataKey = 'userDetail';
+var checkUserName = true;
 
 window.onload = display_data_in_table;
 
-function onKeyDown(event){
+onKeyDown = (event) => {
 	if(event.code == 'Enter'){
-		addPersonData();
+		addDataToStorage();
 	}
-	console.log(event);
 }
 
 // check for the existing data
-function checkLocalStorage(){
+checkLocalStorage = () => {
 	let arr = [];
 	if(localStorage.getItem(dataKey)){
 		arr = JSON.parse(localStorage.getItem(dataKey));
@@ -27,15 +27,14 @@ function checkLocalStorage(){
 
 // when user enter the data validate it and then append the data if there is already some data in localstorage  otherwise simpley add it
 // when data add sucessfully display the data in table forment
-function addDataToStorage() {
+addDataToStorage = () => {
 	createdAt = new Date().toLocaleDateString('en-GB', {day: 'numeric', month: 'short', year: 'numeric'}).replace(/ /g, '-');
-	var checkUserName = false;
+	checkUserName = false;
 
 	firstName = document.getElementById('firstNameId').value;
 	lastName = document.getElementById('lastNameId').value;
 	userName = document.getElementById('userNameId').value;
 	age = document.getElementById('ageId').value;
-
 	userArray = checkLocalStorage();
 
 	var id = document.getElementById('add_data_buton_Id').value;
@@ -53,80 +52,46 @@ function addDataToStorage() {
 		alert('Please Enter Age')
 	}
 	else {
-		for(let i=0; i<userArray.length; i++){
-			checkUserName = userArray[i].userName == userName;
-			if(checkUserName){
-				break;
-			}
-		}
+		var currentUserIndex = userArray.findIndex((elem)=>{
+			return elem.userId == id;
+		});
+		if(currentUserIndex>-1){
+			var oldUserName= userArray[currentUserIndex].userName;
+			if(userName != oldUserName && userNameValidator()){
+				return;
+			} 
 
-		if(checkUserName){
-			alert('UserName is already exist')
-			return;
-		}
-		if(userArray.length){
-			userId = parseInt(userArray[userArray.length -1].userId) + 1;  
-		}
-		
-		if(id > 0){
-			for(let i=0; i<userArray.length; i++){
-				if(id == userArray[i].userId){
-					userArray[i].firstName = document.getElementById('firstNameId').value;
-					userArray[i].lastName = document.getElementById('lastNameId').value;
-					userArray[i].userName = document.getElementById('userNameId').value;
-					userArray[i].age = document.getElementById('ageId').value;
-				}
+			userArray[currentUserIndex].firstName = document.getElementById('firstNameId').value;
+			userArray[currentUserIndex].lastName = document.getElementById('lastNameId').value;
+			userArray[currentUserIndex].userName = document.getElementById('userNameId').value;
+			userArray[currentUserIndex].age = document.getElementById('ageId').value;
+		}else{
+			if(userNameValidator()){
+				return;
 			}
-			document.getElementById('add_data_buton_Id').innerHTML = 'Add Data';
-			document.getElementById('add_data_buton_Id').value = 0;
-		}
-		else{
+			if(userArray.length){
+				userId = parseInt(userArray[userArray.length -1].userId) + 1;  
+			}
 			user = {userId, firstName, lastName, userName, age, createdAt};
-			userArray.push(user);
-		}
-		
+			userArray.push(user);	
+		}			
 		userArrayString = JSON.stringify(userArray);
 		localStorage.setItem(dataKey, userArrayString);
 		
 		display_data_in_table();
-
-		document.getElementById('firstNameId').value = '';
-		document.getElementById('lastNameId').value = '';
-		document.getElementById('userNameId').value = '';
-		document.getElementById('ageId').value = '';
+		resetValue();
 	}
 }
 
-/*function updateData(uid){
-	var array1 = checkLocalStorage();
-
-	for(let i=0; i<array1.length; i++){
-		if(uid == array1[i].userId){
-			array1[i].firstName = document.getElementById('firstNameId').value;
-			array1[i].lastName = document.getElementById('lastNameId').value;
-			array1[i].userName = document.getElementById('userNameId').value;
-			array1[i].age = document.getElementById('ageId').value;
-		}
-	}
-
-	localStorage.setItem(dataKey,JSON.stringify(array1));
-	
-	display_data_in_table();
-
-	document.getElementById('firstNameId').value = '';
-		document.getElementById('lastNameId').value = '';
-		document.getElementById('userNameId').value = '';
-		document.getElementById('ageId').value = '';
-}*/
-
 function display_data_in_table(){
+	var userData = '';
 	userDetailArray = checkLocalStorage();
 
 	if(userDetailArray.length > 0){
 		let userDetailTable = `<table id ='user_data_table' width=100%>
 							   <caption> User Detail </caption>
 							   <tr>
-							   <th> Id </th>
+							   <th> User Id </th>
 							   <th> First Name </th>
 							   <th> Last Name </th>
 							   <th> User-Name </th>
@@ -136,7 +101,7 @@ function display_data_in_table(){
 							   </tr>`;
 		
 		for(let i=0; i < userDetailArray.length; i++){
-			var userData ;
+			
 			userData += `<tr>
 						<td>${userDetailArray[i].userId}</td>
 						<td>${userDetailArray[i].firstName}</td>
@@ -165,7 +130,7 @@ function display_data_in_table(){
 //get the userid of the row and pass it to update function 
 //at update function, get the data-object using userid and set the data into the input_element 
 //when user press the update button send data to the adddata function and update the data at the received id and display change in table
-function updateDatainStorage(id_of_user){
+updateDatainStorage = (id_of_user) => {
 	var usr_id = id_of_user; 	
 	var array = checkLocalStorage();
 	document.getElementById('add_data_buton_Id').innerHTML = 'Update Data';
@@ -178,17 +143,51 @@ function updateDatainStorage(id_of_user){
 			document.getElementById('ageId').value = array[i].age;
 		}
 	}
-
 	document.getElementById('add_data_buton_Id').value = usr_id;
 }
 
 //when press remove : get the id and remove the object from the array and the change should be displayed in table
-function removeDataFromStorage(id_of_user){
+removeDataFromStorage = (id_of_user) => {
+	resetValue();
+
 	var user_id_update = id_of_user;
 	var array_update = checkLocalStorage();
 
-	const updated_array = array_update.filter(user => user.userId != user_id_update )
-	
-	localStorage.setItem(dataKey, JSON.stringify(updated_array));
+	var currentUserIndex = array_update.findIndex((elem)=>{
+			return elem.userId == user_id_update;
+		});
+	var msg_confirmbox = 'Are you really want to Remove : ' + array_update[currentUserIndex].firstName;
+	var confirm_remove_user = confirm(msg_confirmbox);
+
+	if(confirm_remove_user){
+		var updated_array = array_update.filter(
+			function(user){
+				return user.userId != user_id_update;
+			}
+		)
+		localStorage.setItem(dataKey, JSON.stringify(updated_array));
+	}
 	display_data_in_table();
+}
+
+resetValue = () => {
+	document.getElementById('add_data_buton_Id').innerHTML = 'Add Data';
+	document.getElementById('add_data_buton_Id').value = 0;
+	document.getElementById('firstNameId').value = '';
+	document.getElementById('lastNameId').value = '';
+	document.getElementById('userNameId').value = '';
+	document.getElementById('ageId').value = '';
+}
+
+userNameValidator = () => {
+	for(let i=0; i<userArray.length; i++){
+		checkUserName = userArray[i].userName == userName;
+		if(checkUserName){
+			break;
+		}
+	}		
+	if(checkUserName){
+		alert('UserName is already exist')
+		return true;
+	}
 }
